requirements.txt
acres==0.3.0
africastalking==1.2.9
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.11.1
attrs==25.3.0
autobahn==25.12.2
Automat==25.4.16
billiard==4.2.1
blis==1.2.1
boto3==1.42.89
botocore==1.42.89
cachetools==5.5.2
catalogue==2.0.10
cbor2==5.9.0
celery==5.5.2
certifi==2025.4.26
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
chardet==5.2.0
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.21.1
confection==0.1.5
configobj==5.0.9
configparser==7.2.0
constantly==23.10.4
cryptography==46.0.7
cymem==2.0.11
daphne==4.2.1
datasets==3.6.0
dill==0.3.8
dj-database-url==3.1.2
Django==5.2.1
django-ckeditor==6.7.2
django-cleanup==9.0.0
django-cors-headers==4.7.0
django-countries==7.6.1
django-elasticsearch-dsl==7.4
django-elasticsearch-dsl-drf==0.22.5
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
django-formtools==2.5.1
django-js-asset==3.1.2
django-modeltranslation==0.19.19
django-ngrok==0.1.0
django-nine==0.2.7
django-otp==1.6.0
django-phonenumber-field==8.1.0
django-ratelimit==4.1.0
django-recaptcha==4.1.0
django-redis==5.4.0
django-storages==1.14.6
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.2
drf-yasg==1.21.10
ecdsa==0.19.1
elastic-transport==8.17.1
elasticsearch==7.17.12
elasticsearch-dsl==7.4.1
etelemetry==0.3.1
factory-boy==3.3.0
Faker==37.3.0
filelock==3.18.0
fitz==0.0.1.dev2
frozenlist==1.6.0
fsspec==2025.3.0
GDAL==3.8.4
geographiclib==2.0
geoip2==5.2.0
geopandas==1.0.1
geopy==2.4.1
google-auth==2.40.3
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.32.0
hyperlink==21.0.0
idna==3.10
Incremental==24.11.0
inflection==0.5.1
iniconfig==2.1.0
isodate==0.6.1
Jinja2==3.1.6
jmespath==1.1.0
joblib==1.5.1
kombu==5.5.3
langcodes==3.5.0
language_data==1.3.0
looseversion==1.3.0
lxml==5.4.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
maxminddb==3.0.0
mdurl==0.1.2
model-bakery==1.20.4
msgpack==1.1.2
multidict==6.4.4
multiprocess==0.70.16
murmurhash==1.0.13
networkx==3.4.2
ngrok==1.7.0
nibabel==5.3.2
nipype==1.10.0
numpy==1.26.4
onesignal-sdk==2.0.0
packaging==25.0
pandas==2.2.3
pathlib==1.0.1
pdfminer.six==20250327
pdfplumber==0.11.6
phonenumbers==9.0.5
pillow==10.3.0
pluggy==1.6.0
preshed==3.0.9
prompt_toolkit==3.0.51
propcache==0.3.1
prov==2.0.1
psycopg2-binary==2.9.9
puremagic==1.29
py-ubjson==0.16.1
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydot==4.0.0
pyfcm==2.0.8
Pygments==2.19.1
PyJWT==2.10.1
pyogrio==0.11.0
pyOpenSSL==26.0.0
pyotp==2.9.0
pyparsing==3.2.3
pypdfium2==4.30.1
pypng==0.20220715.0
pyproj==3.7.1
pytest==8.3.5
pytest-django==4.11.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-http-client==3.3.7
python-slugify==8.0.4
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
qrcode==7.4.2
rdflib==6.3.2
redis==6.2.0
reportlab==4.4.1
requests==2.32.3
rich==14.0.0
rsa==4.9.1
s3transfer==0.16.0
schema==0.7.7
scikit-learn==1.6.1
scipy==1.15.3
sendgrid==6.12.3
sentry-sdk==2.37.1
service-identity==24.2.0
setuptools==80.8.0
shapely==2.1.1
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
text-unidecode==1.3
thinc==8.3.4
threadpoolctl==3.6.0
tqdm==4.67.1
traits==7.0.2
twilio==9.6.2
Twisted==25.5.0
txaio==25.12.2
typer==0.15.4
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
ujson==5.12.0
uritemplate==4.1.1
urllib3==1.26.20
vine==5.1.0
vonage==4.4.3
vonage-account==1.1.1
vonage-application==2.0.1
vonage-http-client==1.5.1
vonage-jwt==1.1.5
vonage-messages==1.4.0
vonage-network-auth==1.0.2
vonage-network-number-verification==1.0.2
vonage-network-sim-swap==1.1.2
vonage-number-insight==1.0.7
vonage-numbers==1.0.4
vonage-sms==1.1.6
vonage-subaccounts==1.0.4
vonage-users==1.2.1
vonage-utils==1.1.4
vonage-verify==2.1.0
vonage-verify-legacy==1.0.1
vonage-video==1.2.0
vonage-voice==1.4.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
Werkzeug==3.1.3
whitenoise==6.9.0
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.0
zope.interface==8.3
